proper testing in a better way now.
